--- a/SCClassLibrary/Common/Streams/Patterns.sc
+++ b/SCClassLibrary/Common/Streams/Patterns.sc
@@ -182,7 +182,7 @@ Pfuncn : Pattern {
 	storeArgs { ^[func,repeats] }
 	embedInStream {  arg inval;
 		repeats.value(inval).do({
-			inval = func.value(inval).processRest(inval).yield;
+			inval = func.value(inval).yield;
 		});
 		^inval
 	}
@@ -369,7 +369,7 @@ Pbind : Pattern {
 						^inevent
 					};
 					name.do { arg key, i;
-						event.put(key, streamout[i].processRest(event));
+						event.put(key, streamout[i]);
 					};
 				}{
 					event.put(name, streamout);