--- a/SCClassLibrary/Common/Streams/Rest.sc
+++ b/SCClassLibrary/Common/Streams/Rest.sc
@@ -10,6 +10,19 @@ Rest : Operand  {
 	playAndDelta { ^value ? 1 }
 	isRest { ^true }
 
+	prescribeRest { |event|
+		event.put(\isRest, true);
+		if(value.notNil) {
+			event.put(\dur, value.(event))
+		}
+	}
+
+
+}
+
++ Object {
+
+	prescribeRest {}
 
 }