Improve upon the current Rest implementation for event streams.