@telephon 2017-03-10
… we should continue to work on this. @jamshark70 what do you think?