@jamshark70 2017-03-10
I'll merge it into a test branch today and kick the tires for a few days. My live coding framework uses rests in a couple of ways -- if there are any problems, my normal usage pattern will shake them out.

I have a few private tweaks to Rest; I'll also check how they fit the new design.