@telephon 2017-03-10
great ! if we can get this right we can get rid of a lot of complication