@jamshark70 2017-03-11
First try: After resolving some conflicts with my Quarks (which extend Rest and assume its variable is called `dur` rather than the new `value`), the interpreter crashes during the startup sequence. I am rebuilding the SC executable, just to try this class library change. I am surprised it is so invasive.