@jamshark70 2017-03-11
Oh... I see. It's because of a totally unrelated change:

```
Done : UGen {
	// aliases for UGen doneActions
	const <none = 0;
...
```

You know, I remember thinking it was a bad idea to force a new binary just to handle somebody's pet name for a specific value... but I was overruled, and now I have to upgrade everything in order to test this branch, instead of working in my stable sandbox and avoiding the risk of unwanted updates.

Sorry... yeah, my claws are still out. I'm just really really tired of simple things being damn hard.