@jamshark70 2017-03-11
Two practical findings:

1. Is it necessary to change `processRest` to `prescribeRest`? This is forcing changes elsewhere in my quarks.

2. `_Event_Delta` needs a bit of work yet.

```
ERROR: Primitive '_Event_Delta' failed.
Wrong type.
RECEIVER:
Instance of Event {    (0x4915748, gc=6C, fmt=00, flg=00, set=03)
  instance variables [5]
    array : instance of Array (0x7fc3e01a7fa8, size=64, set=6)
    size : Integer 12
    proto : nil
    parent : instance of Event (0x47fe3a8, size=5, set=3)
    know : true
}
PATH: /home/dlm/share/SC/scd/tests/voicer-artic.scd

PROTECTED CALL STACK:
	Meta_MethodError:new	0x7ea6180
		arg this = PrimitiveFailedError
		arg what = Wrong type.
		arg receiver = ( 'clock': a TempoClock, 'eventKey': macroRh, 'dur': Rest(0.25), 'allKeysToPropagate': IdentitySet[ immediateOSC, mode, timingOffset, argKeys, midi, voicer ], 
  'initialRest': true, 'collIndex': ch, 'child': a Proto, 'voicer': a Voicer, 'legato': 0.9, 
  'timingOffset': 0, 'note': Rest(0), 'propagateDownward': a Function )
	Meta_PrimitiveFailedError:new	0x7eac6c0
		arg this = PrimitiveFailedError
		arg receiver = ( 'clock': a TempoClock, 'eventKey': macroRh, 'dur': Rest(0.25), 'allKeysToPropagate': IdentitySet[ immediateOSC, mode, timingOffset, argKeys, midi, voicer ], 
  'initialRest': true, 'collIndex': ch, 'child': a Proto, 'voicer': a Voicer, 'legato': 0.9, 
  'timingOffset': 0, 'note': Rest(0), 'propagateDownward': a Function )
	Object:primitiveFailed	0x7b8d440
		arg this = ( 'clock': a TempoClock, 'eventKey': macroRh, 'dur': Rest(0.25), 'allKeysToPropagate': IdentitySet[ immediateOSC, mode, timingOffset, argKeys, midi, voicer ], 
  'initialRest': true, 'collIndex': ch, 'child': a Proto, 'voicer': a Voicer, 'legato': 0.9, 
  'timingOffset': 0, 'note': Rest(0), 'propagateDownward': a Function )
	a FunctionDef	0x89333c0
		sourceCode = "<an open Function>"
	Integer:do	0x93da280
		arg this = 1
		arg function = a Function
		var i = 0
	PstepDurPair:embedInStream	0x8932f40
		arg this = a PstepDurPair
		arg inval = ( 'clock': a TempoClock, 'eventKey': macroRh, 'dur': Rest(0.25), 'allKeysToPropagate': IdentitySet[ immediateOSC, mode, timingOffset, argKeys, midi, voicer ], 
  'initialRest': true, 'collIndex': ch, 'child': a Proto, 'voicer': a Voicer, 'legato': 0.9, 
  'timingOffset': 0, 'note': Rest(0), 'propagateDownward': a Function )
		var itemStream = a FuncStream
		var durStream = nil
		var pair = [ -, 4 ]
		var item = -
		var dur = 4
		var nextChange = 4
		var elapsed = 0
	a FunctionDef	0x9172d80
		sourceCode = "<an open Function>"
	a FunctionDef	0x9174980
		sourceCode = "<an open Function>"
	Function:prTry	0x8621fc0
		arg this = a Function
		var result = nil
		var thread = a Routine
		var next = nil
		var wasInProtectedFunc = true
	
CALL STACK:
	MethodError:reportError   0xb5ef488
		arg this = <instance of PrimitiveFailedError>
	Nil:handleError   0xb5f0ce8
		arg this = nil
		arg error = <instance of PrimitiveFailedError>
	Thread:handleError   0xb5affd8
		arg this = <instance of Thread>
		arg error = <instance of PrimitiveFailedError>
	Thread:handleError   0xb5d5b38
		arg this = <instance of Routine>
		arg error = <instance of PrimitiveFailedError>
	Object:throw   0xb5eb608
		arg this = <instance of PrimitiveFailedError>
	Function:protect   0xa8d7888
		arg this = <instance of Function>
		arg handler = <instance of Function>
		var result = <instance of PrimitiveFailedError>
	Proto:use   0x61e9dd8
		arg this = <instance of Proto>
		arg func = <instance of Function>
		arg failSilentlyAfterFree = false
		var result = nil
		var saveEnvir = <instance of Environment>
	EventStreamPlayer:prNext   0x61dc068
		arg this = <instance of BlockableEventStreamPlayer>
		arg inTime = 268
		var nextTime = nil
		var outEvent = nil
	< FunctionDef in Method EventStreamPlayer:init >   0xab61a78
		arg inTime = 268
	Routine:prStart   0xab2d0d8
		arg this = <instance of Routine>
		arg inval = 256
^^ The preceding error dump is for ERROR: Primitive '_Event_Delta' failed.
Wrong type.
RECEIVER: ( 'clock': a TempoClock, 'eventKey': macroRh, 'dur': Rest(0.25), 'allKeysToPropagate': IdentitySet[ immediateOSC, mode, timingOffset, argKeys, midi, voicer ], 
  'initialRest': true, 'collIndex': ch, 'child': a Proto, 'voicer': a Voicer, 'legato': 0.9, 
  'timingOffset': 0, 'note': Rest(0), 'propagateDownward': a Function )
```