@telephon 2017-03-11
Hi james, thank you for testing.

1. I wanted to avoid subtly changing the semantics of an existing message, so I used a different name. You are right, that makes it hard to test. But maybe an alias to `processRest` is ok. Could you take a quick look at `prescribeRest` and check if you see any issues? (pasted below)

2. I faintly remember I had changed the primitive, probably even in some PR, but forgotten now. Maybe the primitive could check for a message like `isDelta` instead of for a class?

i don't know if you can get around these issues now.





`
prescribeRest { |event|
	event.put(\isRest, true);
	if(value.notNil) {
		event.put(\dur, value.(event))
	}
}
`