@jamshark70 2017-03-11
Oh, ouch, I didn't notice that `if` branch in `Rest:prescribeRest`. As currently implemented, it assumes that a Rest with a value will only ever be used in `\dur`. Consider:

```
p = Pbind(
    \dur, 0.25,
    \freq, Pseq([Pexprand(200.0, 400.0, 5), Rest(0)], inf)
).play;
```

`prescribeRest` will be called when processing \freq, but the 0 will overwrite \dur. I don't agree with that.

I do have situations in my live coding system where I need Rests in arbitrary keys to carry data. Making it a blanket requirement that a rest's `value` must be nil for every key except `dur` is excessively limiting.

Also it won't help if the pattern bypasses `\dur` and sets `\delta` directly -- the dereferenced value will go into `dur` but the Rest object itself remains in `delta` where it will break rescheduling.

I'd suggest removing the `if` entirely (keep it simple) and properly solve the delta problems where deltas are used.