@jamshark70 2017-03-11
And, FWIW, I thought about it some more and now I think `prescribeRest` is not a replacement for `processRest` -- it's an alternate approach -- but possibly we should formally deprecate it because it may be used in extension classes. (Well, it *is* already used in extension classes...)

So let's keep `prescribeRest`.

And... are we going to stop supporting the Rest class as a data-less placeholder? I'm willing to bet there's user code out there that depends on this.

```
// note: not "a Rest," but "Rest" as a class
Pseq([Pexprand(200.0, 400.0, 5), Rest], inf)
```