@jamshark70 2017-03-18
As expected -- after handling `prEvent_Delta`, I had learned a few tricks to write a new `prEvent_IsRest`. It works as expected (!), and I think it might be the way to go.

Since the performance of scanning the event was a concern, I did some benchmarking with 16 tests:

- `Pbind(\dur, Rest(1))` vs `Pbind(\dur, 1)`
- Two prototype event sizes (5, representing a modestly-sized Pbind, and 100 for an unusually large event)
- Four library code variants:
  1. OLD isRest
  2. NEW isRest: scanning the event in the new primitive, and keeping the old implementation as a fallback for backward compatibility (but I removed the test `this[\isRest] == true` because the primitive replaces it)
  3. NEW isRest, and removing the call to `prescribeRest` from Pbind (there's no need to test `this[\isRest]` anymore, so we can save a bit of time by not setting it at all)
  4. NEW isRest, minus `prescribeRest`, and dropping support for ~freq symbol rests (because that test requires `use`-ing the event, and `use` has an exception handler, and that's a performance killer)

Raw results below. Main conclusions:

- I did the test this way to come closer to the "real world" of using Events with Pbind (and to measure the effect of `prescribeRest` and the `isRest` flag). As a result, we can't really compare e(5) against e(100) because it costs more to copy the larger event prototype. BUT I did a separate test isolating just the primitive, and the performance impact of the event size alone on the primitive is negligible, e(100) / e(5) ~= 5/4.

- The test with the primitive generally performs better than the current way. (I can't explain why non-resting variant 3 was slower than variants 1 or 2, when you'd expect it to be faster. Possibly an outlier.)

- Getting rid of `prescribeRest` might be valuable to simplify pattern internals, but it doesn't save much time. (Getting rid of it would also break compatibility with any user code that expects to look at `event[\isRest]` -- but we could argue that users should have been programming to the public interface `.isRest` instead of to the implementation detail of a collection member.)

- The normal case (non-resting event) is a lot slower than the abnormal case of a rest. This is mostly because of the expensive check for frequency being a Symbol starting with 'r' -- for a six-item event, this one test blows up the execution time by about a factor of 2.5. We could save quite a bit of time by dropping support for the original rest interface -- but that would break compatibility with old code, so I won't argue for that *now*. Worth considering for SC4.

```
// quick test whether the new primitive detects a rest
// even if prescribeRest is *not* used in Pbind
e = Pbind(\dur, Rest(1)).asStream.next(());
e.isRest
-> true  // OK!

// this is the setup for all the e(5) tests below
e = Event.new;
5.do { |i| e.put(("key" ++ i).asSymbol, i) };

// this is the setup for all the e(100) tests below
e = Event.new;
100.do { |i| e.put(("key" ++ i).asSymbol, i) };

// REST cases
(
p = Pbind(\dur, Rest(1)).asStream;
bench { 100000.do { p.next(e.copy).isRest } }
)
OLD
  e(5): time to run: 0.504844141 seconds.
e(100): time to run: 1.1892575 seconds.

NEW with prescribeRest
  e(5): time to run: 0.449129855 seconds.
e(100): time to run: 1.192272353 seconds.

NEW without prescribeRest
  e(5): time to run: 0.41727828199998 seconds.
e(100): time to run: 1.163139545 seconds.

NEW without prescribeRest, and simple fallback
  e(5): time to run: 0.43143614100006 seconds.
e(100): time to run: 1.18602936 seconds.


// NON-REST cases
(
p = Pbind(\dur, 1).asStream;
bench { 100000.do { p.next(e.copy).isRest } }
)
OLD
  e(5): time to run: 1.277349024 seconds.
e(100): time to run: 1.888804284 seconds.

NEW with prescribeRest
  e(5): time to run: 1.272222988 seconds.
e(100): time to run: 1.980391219 seconds.

NEW without prescribeRest
  e(5): time to run: 1.292181877 seconds.
e(100): time to run: 1.975908026 seconds.

NEW without prescribeRest, and simple fallback
  e(5): time to run: 0.48959564400002 seconds.
e(100): time to run: 1.2624454960001 seconds.
```